requirements.txt
dotenv
bcrypt==3.1.4
cython==0.28.3
Flask==1.0.2
Flask-Cors==3.0.4
Flask-Environments==0.1
Flask-Mail==0.9.1
Flask-PyMongo==0.5.1
Flask-Session==0.3.1
intelmq==1.0.4
ipcalc==1.99.0
itsdangerous==0.24
mongoengine==0.15.0
numpy==1.14.3
pandas==0.22.0
Pillow>=5.1.0
pymongo==3.6.1
python-dateutil==2.6.1
python-iptables==0.13.0
pytz==2017.2
redis==2.10.6
regex==2018.2.21
requests==2.18.4
scipy==1.0.0
unicode==2.6
Werkzeug==0.14.1
netifaces
